+++ CHANGELOG.md
@@ -30,7 +30,7 @@
 - `at-rule-no-vendor-prefix`, `media-feature-name-no-vendor-prefix`, `property-no-vendor-prefix`, `selector-no-vendor-prefix`, `value-no-vendor-prefix` to encourage Autoprefixer usage
 - `string-quotes`, `selector-attribute-quotes`, `font-family-name-quotes` to standardise string usage
 - `unspecified: 'bottomAlphabetical'` to `order/properties-order` rule
-
++
 ## 0.1.2
 ### Changed
 - `scss/at-mixin-pattern` changed to include some valid edge cases